generalize and nicify
